+++ trunk/Libraries/Xtro.MDX/ConstantsD3DX10.h
@@ -4,5 +4,5 @@ public:
 	literal unsigned int Default = D3DX10_DEFAULT;
 	literal unsigned int FromFile = D3DX10_FROM_FILE;
 	literal Format FormatFromFile = (Format)DXGI_FORMAT_FROM_FILE;
-	literal float PI = D3DX_PI;
+	literal double PI = D3DX_PI;
 };
